Only fix in async; mark as unsafe
