+++ crates/ruff_linter/src/rules/flake8_trio/mod.rs
@@ -1,4 +1,5 @@
 //! Rules from [flake8-trio](https://pypi.org/project/flake8-trio/).
+pub(super) mod method_name;
 pub(crate) mod rules;
 
 #[cfg(test)]
@@ -14,7 +15,7 @@ mod tests {
     use crate::test::test_path;
 
     #[test_case(Rule::TrioTimeoutWithoutAwait, Path::new("TRIO100.py"))]
-    #[test_case(Rule::SyncTrioCall, Path::new("TRIO105.py"))]
+    #[test_case(Rule::TrioSyncCall, Path::new("TRIO105.py"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.noqa_code(), path.to_string_lossy());
         let diagnostics = test_path(
