+++ crates/ruff_linter/src/rules/flake8_trio/rules/mod.rs
@@ -1,5 +1,5 @@
-pub(crate) use sync_trio_call::*;
+pub(crate) use sync_call::*;
 pub(crate) use timeout_without_await::*;
 
-mod sync_trio_call;
+mod sync_call;
 mod timeout_without_await;
