+++ crates/ruff_linter/src/rules/flake8_trio/rules/timeout_without_await.rs
@@ -1,10 +1,11 @@
 use ruff_diagnostics::{Diagnostic, Violation};
 use ruff_macros::{derive_message_formats, violation};
-use ruff_python_ast::call_path::CallPath;
-use ruff_python_ast::visitor::{walk_expr, walk_stmt, Visitor};
-use ruff_python_ast::{Expr, ExprAwait, Stmt, StmtWith, WithItem};
+use ruff_python_ast::helpers::AwaitVisitor;
+use ruff_python_ast::visitor::Visitor;
+use ruff_python_ast::{StmtWith, WithItem};
 
 use crate::checkers::ast::Checker;
+use crate::rules::flake8_trio::method_name::MethodName;
 
 /// ## What it does
 /// Checks for trio functions that should contain await but don't.
@@ -56,6 +57,17 @@ pub(crate) fn timeout_without_await(
         return;
     };
 
+    if !matches!(
+        method_name,
+        MethodName::MoveOnAfter
+            | MethodName::MoveOnAt
+            | MethodName::FailAfter
+            | MethodName::FailAt
+            | MethodName::CancelScope
+    ) {
+        return;
+    }
+
     let mut visitor = AwaitVisitor::default();
     visitor.visit_body(&with_stmt.body);
     if visitor.seen_await {
@@ -66,60 +78,4 @@ pub(crate) fn timeout_without_await(
         TrioTimeoutWithoutAwait { method_name },
         with_stmt.range,
     ));
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum MethodName {
-    MoveOnAfter,
-    MoveOnAt,
-    FailAfter,
-    FailAt,
-    CancelScope,
-}
-
-impl MethodName {
-    fn try_from(call_path: &CallPath<'_>) -> Option<Self> {
-        match call_path.as_slice() {
-            ["trio", "move_on_after"] => Some(Self::MoveOnAfter),
-            ["trio", "move_on_at"] => Some(Self::MoveOnAt),
-            ["trio", "fail_after"] => Some(Self::FailAfter),
-            ["trio", "fail_at"] => Some(Self::FailAt),
-            ["trio", "CancelScope"] => Some(Self::CancelScope),
-            _ => None,
-        }
-    }
-}
-
-impl std::fmt::Display for MethodName {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            MethodName::MoveOnAfter => write!(f, "trio.move_on_after"),
-            MethodName::MoveOnAt => write!(f, "trio.move_on_at"),
-            MethodName::FailAfter => write!(f, "trio.fail_after"),
-            MethodName::FailAt => write!(f, "trio.fail_at"),
-            MethodName::CancelScope => write!(f, "trio.CancelScope"),
-        }
-    }
-}
-
-#[derive(Debug, Default)]
-pub(crate) struct AwaitVisitor {
-    pub(crate) seen_await: bool,
-}
-
-impl Visitor<'_> for AwaitVisitor {
-    fn visit_stmt(&mut self, stmt: &Stmt) {
-        match stmt {
-            Stmt::FunctionDef(_) | Stmt::ClassDef(_) => (),
-            _ => walk_stmt(self, stmt),
-        }
-    }
-
-    fn visit_expr(&mut self, expr: &Expr) {
-        if let Expr::Await(ExprAwait { .. }) = expr {
-            self.seen_await = true;
-        } else {
-            walk_expr(self, expr);
-        }
-    }
 }
